requirements.txt
fastapi
uvicorn
supabase
ollama
duckduckgo-search
python-dotenv
python-multipart